[adapters] move JobQueue and backoff into adapterlib
The Iceberg crate can't depend on dbsp_adapters (adapters already depends
on feldera-iceberg, so the reverse edge would be a cycle). Move JobQueue
and calculate_backoff_delay into feldera_adapterlib::utils so the Delta
and Iceberg connectors can share them.

No behavior change: the Delta connector re-imports both from adapterlib.
