+++ crates/adapterlib/Cargo.toml
@@ -39,7 +39,7 @@ rmpv = { workspace = true, features = ["with-serde"] }
 bytemuck = { workspace = true }
 num-traits = { workspace = true }
 num-derive = { workspace = true }
-tokio = { workspace = true, features = ["sync"] }
+tokio = { workspace = true, features = ["sync", "rt"] }
 xxhash-rust = { workspace = true, features = ["xxh3"] }
 datafusion = { workspace = true }
 datafusion-functions-json = { workspace = true }
@@ -47,6 +47,9 @@ thiserror = { workspace = true }
 serde_json_path_to_error = { workspace = true }
 chrono = { workspace = true }
 tracing = { workspace = true }
+async-channel = { workspace = true }
+futures = { workspace = true }
+rand = { workspace = true }
 
 [package.metadata.cargo-machete]
 ignored = ["num-traits"]
