+++ crates/adapterlib/src/utils.rs
@@ -1,1 +1,3 @@
+pub mod backoff;
 pub mod datafusion;
+pub mod job_queue;
